+++ calculator/src/main/java/calculator/CalcModel.java
@@ -1,0 +1,45 @@
+package calculator;
+
+/**
+ * A state class which switches between postfix and infix.
+ *
+ * @author HIJI
+ */
+public class CalcModel {
+  //The state that will be tracked
+  private Calculator currentState;
+  //Possible states for the calculator
+  private Calculator scalc;
+  private Calculator rcalc;
+  
+  /**
+   * Initially sets the current state of the model to {@link calculator.RevPolishCalc
+   * RevPolishCalc}.
+   */
+  public CalcModel() {
+    scalc = new StandardCalc();
+    rcalc = new RevPolishCalc();
+    currentState = rcalc;
+  }
+  
+  /**
+   * Change calculator to either infix or postfix.
+   *
+   * @param type The mode calculator should be in 
+   */
+  public void setState(OpType type) {
+    //Check if state should be in infix mode
+    if (type == OpType.INFIX) {
+      currentState = scalc;
+    } else {
+      currentState = rcalc;
+    }
+  }
+  
+  /**
+   * Evaluates user expression either in infix or postfix.
+   */
+  public float evaluate(String expression) {
+    return currentState.evaluate(expression);
+  }
+}