+++ calculator/src/main/java/calculator/GuiView.java
@@ -1,10 +1,16 @@
 package calculator;
 
+import java.util.function.Consumer;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.control.RadioButton;
 import javafx.scene.control.TextField;
+import javafx.scene.control.Toggle;
+import javafx.scene.control.ToggleGroup;
 import javafx.scene.text.Text;
 
 /**
@@ -48,6 +54,20 @@ public class GuiView implements ViewInterface {
   }
   
   /**
+   * This allows a new observer to be added to the change calculate strategy action.
+   *
+   * @param l the Observer to be notified.
+   */
+  public void addTypeObserver(Consumer<OpType> l) {
+    calcType.selectedToggleProperty().addListener(new ChangeListener<Toggle>() {
+      @Override
+      public void changed(ObservableValue<? extends Toggle> observable, Toggle from, Toggle to) {
+        l.accept(to == infixButton ? OpType.INFIX : OpType.POSTFIX);
+      }
+    });
+  }
+  
+  /**
    * The text object which displays the answer on the screen.
    */
   @FXML
@@ -67,4 +87,25 @@ public class GuiView implements ViewInterface {
   //fx:id="submit"
   @FXML
   private Button submit; // Value injected by FXMLLoader
+  
+  /**
+   * The polish radio button on the application.
+   */
+  //fx:id="polishButton"
+  @FXML
+  private RadioButton polishButton;
+  
+  /**
+   * The infix radio button on the application.
+   */
+  //fx:id="infixButton"
+  @FXML
+  private RadioButton infixButton;
+  
+  /**
+   * The object that links the two radio buttons.
+   */
+  //fx:id="calcType"
+  @FXML
+  private ToggleGroup calcType;
 }